% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_synonyms.R
\name{bd_synonym}
\alias{bd_synonym}
\title{Function to get synonyms of the species.}
\usage{
bd_synonym(name = NULL)
}
\arguments{
\item{name}{(string) scientfic name of the species.}
}
\value{
returns a data.frame which contains synonyms and other related taxonomic info.
}
\description{
Function to get synonyms of the species.
}
