% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_taxonomic_resolver.R
\name{bd_taxonomic_resolver}
\alias{bd_taxonomic_resolver}
\title{Function to resolve biodiveristy data based on taxonomic ranks.}
\usage{
bd_taxonomic_resolver(df = NULL)
}
\arguments{
\item{df}{(data.frame) Data frame containing biodiversity data.}
}
\value{
(data.frame) Returns a data frame containing resolved biodiversity data.
}
\description{
Function to resolve biodiveristy data based on taxonomic ranks.
}
