% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_taxonomic_validation.R
\name{bd_taxonomic_validation}
\alias{bd_taxonomic_validation}
\title{Function for taxonomic validation}
\usage{
bd_taxonomic_validation(df = NULL)
}
\arguments{
\item{df}{(data.frame) contains a data frame with species data}
}
\value{
(data.frame) returns a data frame containing filtered data based on user choice
}
\description{
Function for taxonomic validation
}
